fix: Full lint + test hardening — 486 ruff errors to zero, 10 test failures to zero
Pytest burn-down (10 failures -> 0):
- Fix test_system_health: use importlib.reload for module reimport
- Fix test_production_index_baseline: correct OrderStatus enum and position mocks
- Fix test_api_managers: update assertions for numeric return type
- Fix test_market_regime_adapter: align with neutral default indicators

Ruff lint burn-down (486 -> 0):
- B904: Add exception chaining (from e) across 11 files
- RUF012: Annotate 49 mutable class defaults with ClassVar across 30 files
- F821: Add missing imports (datetime, timezone, numpy, pandas) in 3 files
- E741: Rename ambiguous variables (l -> lot/level/leg/low) in 7 files
- F811: Remove duplicate function definitions, rename shadowed test class
- RUF034: Remove useless if-else in renko.py, message_types.py, telegram_bot.py
- RUF002/RUF003: Replace unicode math symbols with ASCII in docstrings/comments
- B017: Narrow pytest.raises(Exception) to specific types (ValueError, TypeError, IntegrityError)
- S106/S301/S324/S104/ASYNC240: Add targeted noqa for intentional security patterns
- F823: Remove redundant local timedelta imports shadowing module-level import
- PLW2901: Avoid loop variable reassignment in trade_reasoning.py
- RUF006: Store asyncio.create_task reference on self to prevent GC
- Fix unsafe ruff autofix: callable | None -> object | None in dashboard_service.py

+++ backend/auth0login/services/risk_assessment.py
@@ -197,7 +197,7 @@ class RiskAssessmentService:
                     break
 
         # Round to integer for simplicity
-        final_score = int(round(total_weighted_score))
+        final_score = round(total_weighted_score)
 
         # Determine profile
         recommended_profile = 'moderate'  # default
@@ -208,7 +208,7 @@ class RiskAssessmentService:
 
         return QuestionnaireResult(
             total_score=final_score,
-            max_possible_score=int(round(max_weighted_score)),
+            max_possible_score=round(max_weighted_score),
             recommended_profile=recommended_profile,
             profile_explanation=PROFILE_EXPLANATIONS[recommended_profile],
             score_breakdown=score_breakdown,
